rename `EntityConfigState` to `GenderRenderState`, move to `render` package
