+++ src/main/java/com/wildfire/render/GenderEntityRenderState.java
@@ -19,14 +19,13 @@
 package com.wildfire.render;
 
 import com.wildfire.main.entitydata.EntityConfig;
-import com.wildfire.main.entitydata.EntityConfigState;
 import net.minecraft.text.Text;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
 // See LivingEntityRenderStateMixin for the actual implementation
 public interface GenderEntityRenderState {
-	@NotNull EntityConfigState getEntityConfigState();
+	@NotNull GenderRenderState getEntityConfigState();
 	void updateEntityConfigState(EntityConfig entityConfig);
 
 	boolean isBreathing();
